Restaurant activity

See merge request kswidere/pap22Z-z23!6

+++ app/src/main/java/pw/pap22z/bulionapp/data/entities/Review.kt
@@ -10,7 +10,7 @@ import kotlinx.parcelize.Parcelize
 @Entity
 data class Review(
     @PrimaryKey(autoGenerate = true) val review_id: Int = 0,
-    val review_rating: Int,
+    val review_rating: Float,
     val review_body: String,
     @Embedded val restaurant: Restaurant?,
     @Embedded val user: User
